Use "C-M-j" and "C-M-k" for tag jumps
Example 1: start with "C-M-k" to find something in the current file.
It doesn't exist, to press "C-M-j" to search for the /same input/ in the
whole directory.

Example 2: start with "C-M-j" and get to many cands. Press "C-M-k" to
narrow the candidates to the file of the current candidate.

* function-args.el (map): Update keymap.
(moo-complete--candidates): New defun.
(moo-complete): Update.
(moo-completion-at-point): New defun.
(moo-jump-keymap): New keymap.
(moo-jump-local): Now it's really local, get tags for the current file,
bound to "C-M-k".
(moo-jump-directory): Jump to tags in the current dir, bound to "C-M-j".
(moo-select-candidate): Update.
(moo-action-jump): Update.

Re abo-abo/swiper#606